+++ src/RobotStudio.Desktop/Examples/RobotExampleExpectedResult.cs
@@ -1,0 +1,7 @@
+namespace RobotStudio.Desktop.Examples;
+
+public enum RobotExampleExpectedResult
+{
+    Succeeds,
+    ValidationError
+}